Separate pager payload state
