+++ src/Repl.Core/ResultFlow/PagerHeader.cs
@@ -1,0 +1,6 @@
+namespace Repl;
+
+internal sealed record PagerHeader(IReadOnlyList<string> Lines, IReadOnlySet<string> NormalizedLines)
+{
+	public static PagerHeader Empty { get; } = new([], new HashSet<string>(StringComparer.Ordinal));
+}
