+++ src/Repl.Core/ResultFlow/ParsedPagerPayload.cs
@@ -1,0 +1,6 @@
+namespace Repl;
+
+internal sealed record ParsedPagerPayload(PagerHeader Header, IReadOnlyList<string> ContentLines)
+{
+	public int TotalLineCount => Header.Lines.Count + ContentLines.Count;
+}
