+++ src/Repl.Core/ResultFlow/ResultFlowPagerPage.cs
@@ -1,3 +1,6 @@
 namespace Repl;
 
-internal sealed record ResultFlowPagerPage(string Payload, bool HasMore);
+internal sealed record ResultFlowPagerPage(
+	string Payload,
+	bool HasMore,
+	bool ContainsPresentationChrome = true);
